Change tool location to sut_orchestrator
