+++ lisa/sut_orchestrator/azure/tools.py
@@ -380,3 +380,17 @@ class AzCmdlet(Tool):
             f"-CommandId 'RunPowerShellScript' "
             f"-Parameter @{{'PublicKey'='{public_key_data}'}}"
         )
+
+class Azsecd(Tool):
+    @property
+    def command(self) -> str:
+        return "azsecd"
+
+    @property
+    def can_install(self) -> bool:
+        return False
+
+    def install(self) -> bool:
+        package_name = "azure-security"
+        self.node.os.install_packages(package_name)
+        return self._check_exists()
